Extract auto-rotate loop into useAutoRotate hook
